Ensure pointers to CFG_TUSB_MEM_SECTION also have the qualifier
Whatever storage qualifiers are added to data types should be also
added to pointers to those types, to avoid unnecessary conversions (or
outright errors, depending on the qualifier/section). In the case of
Harvard architectures like the 8051 (supported by SDCC) you actually
need to know what kind of pointer it is to generate the correct instructions.