Blinking is working!
It's unclear why but previously (1 month ago) this was too large to
fit in memory according to sdcc. I gave up in disgust and started
messing with the riscv emulator and now *this* works.

Anyway just as in the rv51 branch we use the main flatbed light as a
standin for an LED.