Introduce a new interface for debugging via the interrupt endpoint
The MSC class does not normally use this (only in the CPI protocol)
and so it is available for other purposes. Previously this endpoint
was available for debug messages only until the device was reset,
which causes the descriptors to be re-read. Linux will not allow me to
access an endpoint that is not listed in the descriptors. Now it will
be available post-reset (i.e. during normal operation of the device).